+++ Branch-SDK/Branch-SDK/BNCFieldDefines.h
@@ -101,8 +101,10 @@
                 } \
             } \
             object.field = newArray; \
+        } else if ([a isKindOfClass:[NSString class]]) { \
+            object.field = [NSMutableArray arrayWithObject:a]; \
         } else { \
-            object.field = (id) [NSArray new]; \
+            object.field = (id) [NSMutableArray new]; \
         } \
     }