+++ Branch-TestBed/Branch-SDK-Tests/BranchEvent.Test.swift
@@ -10,6 +10,10 @@ import Foundation
 
 class BranchEventTestSwift : BNCTestCase {
 
+    override func setUp() {
+        Branch.getInstance("key_live_foo")
+    }
+
     func testBranchEvent() {
 
         // Set up the Branch Universal Object --
@@ -40,6 +44,7 @@ class BranchEventTestSwift : BNCTestCase {
         branchUniversalObject.contentMetadata.ratingAverage     = 5;
         branchUniversalObject.contentMetadata.ratingCount       = 5;
         branchUniversalObject.contentMetadata.ratingMax         = 7;
+        branchUniversalObject.contentMetadata.rating            = 6;
         branchUniversalObject.contentMetadata.addressStreet     = "Street_name1"
         branchUniversalObject.contentMetadata.addressCity       = "city1"
         branchUniversalObject.contentMetadata.addressRegion     = "Region1"
@@ -47,9 +52,13 @@ class BranchEventTestSwift : BNCTestCase {
         branchUniversalObject.contentMetadata.addressPostalCode = "postal_code"
         branchUniversalObject.contentMetadata.latitude          = 12.07
         branchUniversalObject.contentMetadata.longitude         = -97.5
-        branchUniversalObject.contentMetadata.imageCaptions     = [ "my_img_caption1",  "my_img_caption_2"]
+        branchUniversalObject.contentMetadata.imageCaptions     = [
+            "my_img_caption1",
+            "my_img_caption_2"
+        ]
         branchUniversalObject.contentMetadata.customMetadata    = [
-            "Custom_Content_metadata_key1": "Custom_Content_metadata_val1"
+            "Custom_Content_metadata_key1": "Custom_Content_metadata_val1",
+            "Custom_Content_metadata_key2": "Custom_Content_metadata_val2"
         ]
 
         // Set up the event properties --
@@ -97,5 +106,16 @@ class BranchEventTestSwift : BNCTestCase {
         event.logEvent()
 
         event = BranchEvent.standardEvent(.viewItem)
+        event.logEvent();
+
+        // Quickly log an event:
+        BranchEvent.standardEvent(.viewItem).logEvent()
+
+        // Quickly log an event with content:
+        let branchUniversalObject = BranchUniversalObject.init()
+        branchUniversalObject.canonicalIdentifier = "item/12345"
+        branchUniversalObject.canonicalUrl        = "https://branch.io/deepviews"
+        branchUniversalObject.title               = "My Content Title"
+        BranchEvent.standardEvent(.viewItem, withContentItem: branchUniversalObject).logEvent()
     }
 }