+++ Branch-TestBed/Branch-TestBed-UITests/UITestBed/TBBranchViewController.m
@@ -14,7 +14,6 @@
 @import Branch;
 #import "BNCDeviceInfo.h"
 #import "BNCSystemObserver.h"
-#import "UIViewController+Branch.h"
 
 NSString *cononicalIdentifier = @"item/12345";
 NSString *canonicalUrl = @"https://dev.branch.io/getting-started/deep-link-routing/guide/ios/";