+++ ChangeLog.md
@@ -1,5 +1,29 @@
 Branch iOS SDK Change Log
 
+- v0.22.5
+  * _*Master Release*_ - January 12, 2018
+  * Added the `rating` field to the `BranchUniversalObject`.
+
+- v0.22.4
+  * _*Master Release*_ - January 3, 2018
+  * Fix for crash when the application delegate does not have a window variable.
+
+- v0.22.3
+  * _*Master Release*_ - December 29, 2017
+  * Fixed the static library build.
+
+- v0.22.2
+  * _*Master Release*_ - December 29, 2017
+
+- v0.22.1
+  * Flattened the Branch Universal Object dictionary to be fully compatible with Android.
+
+- v0.22.0
+  * Apple search ads weren't timing out on iPad during initialization.
+    - This could block Branch initialization when Search Ad attribution was turned on.
+  * Fixed a deadlock found in iOS 9.3 / iPhone 5s.
+  * Added the `[UIViewController bnc_currentViewController]` as a category.
+
 - v0.21.16
   * _*Master Release*_ - December 18, 2017
   * Fixed the BranchUniversalObject for v2-events & updated the UITestBed. (#768)